backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
channels==4.2.0
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
python-decouple==3.8
sqlparse==0.5.3
uvicorn==0.34.0
websockets==14.2
whitenoise==6.8.2
